chore: restructure dirs
